requirements.txt
fastapi
uvicorn[standard]
transformers>=4.30.0
torch
sentence-transformers
langchain
faiss-cpu
pdfplumber
pytesseract
spacy
pandas
presidio-analyzer
presidio-anonymizer
python-multipart
reportlab